requirements.txt
fastapi 
uvicorn 
opencv-python 
numpy
Pillow
scipy 
python-multipart
asyncpg 
psycopg2-binary
databases
pyjwt
pydantic[email]
insightface
onnxruntime
passlib
slowapi
jinja2
pyotp
qrcode
dotenv
boto3
awscli
openai
resend
fastapi[standard]
sqlalchemy
passlib[bcrypt]
alembic